Add files via upload
This is my code for gradient decent from scratch. 
This is for intuition point of view. I have taken two numpy array x and y. As we know equation of straight line is for linearly related data is y = mx+c. where m is slope and c is an intercept.
Here I did approx 1000 trials to get best slope and intercept. and by using that slope and intercept we can form our equation like as y = mx+c and we can predict for future data.
Data:
x = np.array([1,2,3,4,5])
y = np.array([5,7,9,11,13])